+++ index.html
@@ -165,7 +165,7 @@
     </section>
   </template>
 
-  <script src="/js/main.js"></script>
+  <script type="module" src="/js/main.js"></script>
 
 </body>
 </html>
